update: add check for JaCoCo XML report existence in SonarQube analysis stage

+++ Jenkinsfile
@@ -53,22 +53,30 @@ pipeline {
         }
 
           stage('SonarQube Analysis') {
-                    steps {
-                        echo 'Running SonarQube analysis...'
-                        withSonarQubeEnv("${SONARQUBE_ENV}") {
-                            sh """
-                            docker run --rm -v \$(pwd):/app -w /app ${DOCKER_IMAGE} \
-                            mvn sonar:sonar \
-                                -Dsonar.projectKey=psoft-g1-staging \
-                                -Dsonar.projectName="PSoft G1 Project Staging" \
-                                -Dsonar.host.url=http://74.161.33.56:9000 \
-                                -Dsonar.login=squ_186e07b99759c0ff10a3f1127bbb2b79ed20a393 \
-                                -Dsonar.java.coveragePlugin=jacoco \
-                                -Dsonar.coverage.jacoco.xmlReportPaths=target/jacoco-report/jacoco.xml
-                            """
-                        }
-                    }
-                }
+              steps {
+                  echo 'Running SonarQube analysis...'
+                  withSonarQubeEnv("${SONARQUBE_ENV}") {
+                      sh """
+                      # Certifica que o Jacoco XML existe antes
+                      if [ ! -f target/jacoco-report/jacoco.xml ]; then
+                          echo "❌ Jacoco XML not found!"
+                          ls -l target/jacoco-report
+                          exit 1
+                      fi
+
+                      docker run --rm -v \$(pwd):/app -w /app ${DOCKER_IMAGE} \
+                      mvn sonar:sonar \
+                          -Dsonar.projectKey=psoft-g1-staging \
+                          -Dsonar.projectName="PSoft G1 Project Staging" \
+                          -Dsonar.host.url=http://74.161.33.56:9000 \
+                          -Dsonar.login=squ_186e07b99759c0ff10a3f1127bbb2b79ed20a393 \
+                          -Dsonar.java.coveragePlugin=jacoco \
+                          -Dsonar.coverage.jacoco.xmlReportPaths=target/jacoco-report/jacoco.xml
+                      """
+                  }
+              }
+          }
+
         stage('Quality Gate') {
             steps {
                 timeout(time: 2, unit: 'MINUTES') {
